Fold 'hoistModel' into 'XH'
+ Creates a single more-convenient singleton with all API entry points.
+ Improve documentation
