+++ admin/tabs/configs/differ/ConfigDifferModel.js
@@ -65,7 +65,7 @@ export class ConfigDifferModel  {
             const resp = await Promise.all([
                 XH.fetchJson({url: XH.baseUrl + 'configDiffAdmin/configs'}),
                 XH.fetchJson({url: trimEnd(this.remoteHost, '/') + '/configDiffAdmin/configs'})
-            ]).linkTo(XH.hoistModel.appLoadModel);
+            ]).linkTo(XH.appLoadModel);
             this.processResponse(resp);
         } catch (e) {
             this.processFailedLoad();
@@ -165,7 +165,7 @@ export class ConfigDifferModel  {
             this.loadAsync();
             this.detailModel.close();
         }).linkTo(
-            XH.hoistModel.appLoadModel
+            XH.appLoadModel
         ).catchDefault();
     }
