+++ app/impl/LoginPanel.js
@@ -87,7 +87,7 @@ export class LoginPanel extends Component {
         }).then(r => {
             this.setWarning(r.success ? '' : 'Login Incorrect');
             if (r.success) {
-                XH.hoistModel.completeInitAsync(username);
+                XH.completeInitAsync(username);
             }
         }).catchDefault();
     }
